requirements.txt
torch>=2.0.0
transformers>=4.40.0
pillow>=9.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0
